refactor: Move stats parsing logic from UI to Docker package
- Created ContainerStats model in internal/models
- Added ParseStatsJSON function to internal/docker/parser.go
- Implemented GetStats method in docker package using ParseStatsJSON
- Simplified loadStats function to use client.GetStats()

This properly separates concerns - the docker package handles Docker API
interaction and JSON parsing (consistent with ParsePSJSON pattern), while
the UI layer just consumes the data.
